Addresses windows PATH issue preventing some modules from loading dependencies.
If `/env/Library/bin` is not on the Windows PATH, packages are unable to load dependencies in that directory.  Issue is addressed by extending ExecuteCondaEnvironmentCommand with context manager protocols that temporarily prepend the required directory to `os.environ['PATH']` prior to running the target file. The `PATH` entry is reset to its original value once target file exits.

The patch is only applied to Windows environments where conda version <= 4.6