Fixed performance issue affecting Windows users introduced in PR #15
The subprocess call that retrieves the installed conda version (in json format) causes a noticeable hang.  To address this, the version number is now cached after first access.

While a small delay still occurs the first time a file is ran, the problem will not propagate to subsequent calls until restarting the editor or reloading the plugin.